Fixed FPN multiplication
Turned out I had to both change the destination mantissa start point to
be at most dbits - dexp and increase the mantissa end point to borrow
an exponent bit for the assumed bit. There may still be a bug but that's
for more thorough testing later after I get division working since any
bug will mostly be a subtlety or edge case, both of which need basic
functionality working first to find (which is now the case).